# src/docs/swagger.yaml
openapi: 3.1.0
info:
  title: Test task API
  version: 1.0.0
  description: API for Test task
servers:
  - url: http://localhost:7852
paths:
...
  /api/files:
    post:
      summary: File upload
      tags:
        - Files
      operationId: File upload
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  required: true
      responses:
        200:
          description: File upload
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: string
                    example: ok
                  message:
                    type: string
                    example: File upload successfully
    get:
      summary: Files list
      tags:
        - Files
      operationId: Files list
      security:
        - BearerAuth: [ ]
      responses:
        200:
          description: Files list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: string
                    example: ok
                  message:
                    type: string
                    example: Files list
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        user_id:
                          type: string
                        name:
                          type: string
                        mimetype:
                          type: string
                        extension:
                          type: string
                        size:
                          type: string
                        url:
                          type: string
                        created_at:
                          type: string
                          format: date-time
                        updated_at:
                          type: string
                          format: date-time
                    count:
                      type: number
  /api/files/{id}:
    put:
      summary: File upload
      tags:
        - Files
      operationId: File upload
      security:
        - BearerAuth: [ ]
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the file to retrieve
          schema:
            type: string
            format: uuid
          example: "3d2052b7-7744-4d52-94dd-e671d476bb43"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  required: true
      responses:
        200:
          description: File upload
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: string
                    example: ok
                  message:
                    type: string
                    example: File upload successfully
    get:
      summary: File get by id
      tags:
        - Files
      operationId: File get by id
      security:
        - BearerAuth: [ ]
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the file to retrieve
          schema:
            type: string
            format: uuid
          example: "3d2052b7-7744-4d52-94dd-e671d476bb43"
      responses:
        200:
          description: File destroy
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: string
                    example: ok
                  message:
                    type: string
                    example: File
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      user_id:
                        type: string
                      name:
                        type: string
                      mimetype:
                        type: string
                      extension:
                        type: string
                      size:
                        type: string
                      url:
                        type: string
                      created_at:
                        type: string
                        format: date-time
                      updated_at:
                        type: string
                        format: date-time
    delete:
      summary: File destroy
      tags:
        - Files
      operationId: File destroy
      security:
        - BearerAuth: [ ]
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the file to retrieve
          schema:
            type: string
            format: uuid
          example: "3d2052b7-7744-4d52-94dd-e671d476bb43"
      responses:
        200:
          description: File destroy
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: string
                    example: ok
                  message:
                    type: string
                    example: File destroy successfully
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT